requirements.txt
discord.py[voice]
python-dotenv
yt-dlp
aiosqlite